refactor: reuse curtailment helpers
