+++ client/src/protoFleet/features/energy/ActiveCurtailmentStatus.stories.tsx
@@ -1,9 +1,7 @@
 import { type ComponentProps, type ReactElement, useEffect, useMemo, useState } from "react";
 import type { Meta, StoryObj } from "@storybook/react-vite";
 
-import ActiveCurtailmentStatus, {
-  type ActiveCurtailmentEvent,
-} from "@/protoFleet/features/energy/ActiveCurtailmentStatus";
+import ActiveCurtailmentStatus from "@/protoFleet/features/energy/ActiveCurtailmentStatus";
 import {
   curtailedCurtailmentEvent,
   curtailingCurtailmentEvent,
@@ -14,6 +12,7 @@ import {
 import CurtailmentStopConfirmationDialog, {
   type CurtailmentStopConfirmationAction,
 } from "@/protoFleet/features/energy/CurtailmentStopConfirmationDialog";
+import type { ActiveCurtailmentEvent } from "@/protoFleet/features/energy/curtailmentTypes";
 
 const meta = {
   title: "Proto Fleet/Energy/Active Curtailment Status",
