+++ client/src/protoFleet/features/energy/CurtailmentHistory.fixtures.ts
@@ -1,4 +1,4 @@
-import type { CurtailmentHistoryEvent } from "@/protoFleet/features/energy/CurtailmentHistory";
+import type { CurtailmentHistoryEvent } from "@/protoFleet/features/energy/curtailmentTypes";
 
 export const mockCurtailmentHistoryEvents: CurtailmentHistoryEvent[] = [
   {
